agents/butler/requirements.txt
# AWS SDK for Bedrock
boto3==1.35.36
botocore==1.35.36

# PostgreSQL adapter
psycopg2-binary==2.9.9

# Environment variables
python-dotenv==1.0.0

# Utilities
requests==2.31.0